Refactor Dockerfile to copy all source files and rename mesh-monitor.py to src directory
